offering & rfq working, but quote is 404'ing
