--- src/dwn-proxy-markup/protocol-route-action.ts
@@ -1,9 +1,0 @@
-
-
-
-export interface IProtocolRouteAction {
-  id: string
-  description: string
-  action: string
-  params: 
-}
